[FEATURE] function that returns the number of bits + change in the data type

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -1,7 +1,7 @@
 #include "holberton.h"
 
 /**
- * flip_bits
+ * flip_bits - function that returns the number of bits
  * @n: value
  * @m: position
  * Return: to the counter.
@@ -10,7 +10,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	unsigned long int kalkulilo = 0;
-	unsigned long int xor;
+	unsigned int xor;
 
 	xor = n ^ m;
